encapsulate carousel initial data
