+++ src/componentAtoms/CarouselCommon/utils.js
@@ -1,0 +1,29 @@
+// temp data array
+const slides = [
+	{
+		src: require('assets/img/denys.jpg'),
+		altText: 'Slide 1',
+	},
+	{
+		src: require('assets/img/fabien-bazanegue.jpg'),
+		altText: 'Slide 2',
+	},
+	{
+		src: require('assets/img/mark-finn.jpg'),
+		altText: 'Slide 3',
+	},
+	{
+		src: require('assets/img/denys.jpg'),
+		altText: 'Slide 4',
+	},
+	{
+		src: require('assets/img/fabien-bazanegue.jpg'),
+		altText: 'Slide 5',
+	},
+	{
+		src: require('assets/img/mark-finn.jpg'),
+		altText: 'Slide 6',
+	},
+]
+
+export { slides }
